+++ examples/resources/morpheus_max_vms_policy/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_max_vms_policy.tf_example_max_vms_policy 1